Update gap_code.g
During the revision process, we realized that the GAP command "<" does not compare irrational numbers correctly. The original version of this code relied on that command to check some inequalities that may or may not have involved irrational numbers, so there was a possibility for error. This version uses the package FUtil (http://www.math.rwth-aachen.de/~Frank.Luebeck/gap/FUtil/index.html) to make those comparisons instead. As it turns out, the final results are the same as those produced by the original code. In particular, it was not necessary to run the Sage code again.